requirements.txt
fastapi
uvicorn
python-telegram-bot
langchain-openai
langgraph
python-dotenv
httpx
pydantic-settings
pyyaml
structlog
